tools/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
geopy==2.4.0
numpy==1.24.3
tqdm==4.66.1 